+++ docs/src/background_task_lock_controller/background_task_lock_controller.py
@@ -3,7 +3,7 @@ import time
 from apidaora import appdaora, route
 
 
-@route.background('/hello-single', single_running=True)
+@route.background('/hello-single', lock=True)
 def hello_task(name: str) -> str:
     time.sleep(1)
     return f'Hello {name}!'